fix: unblock waste management pr gates
