+++ packages/auth-runtime/src/waste-management/core/tours.ts
@@ -1,12 +1,15 @@
-import { createSdkLogger } from '@sva/server-runtime';
-
 import type { AuthenticatedRequestContext } from '../../middleware.js';
-import { buildLogContext } from '../../log-context.js';
 import { validateCsrf } from '../../shared/request-security.js';
 import { asApiItem, createApiError, parseRequestBody, readPathSegment } from '../../shared/request-helpers.js';
 import { authorizeWasteManagementAction, emitWasteAuditEvent } from './auth.js';
 import { wasteManagementTourSchemas } from './schemas.js';
 import { updateWasteVisibleStatus } from './settings-shared.js';
+import {
+  createWasteTourDeleteErrorResponse,
+  logWasteTourDeleteFinalDelete,
+  logWasteTourDeleteLoaded,
+  logWasteTourDeleteRequested,
+} from './tours-delete-support.js';
 import {
   createWasteManagementTourAfterValidation,
   createWasteTourWriteInput,
@@ -16,29 +19,6 @@ import type { WasteManagementHandlerDeps } from './types.js';
 import { getRequestId, requireActorInstanceId, requireDeps } from './utils.js';
 
 const { createWasteTourSchema, updateWasteTourSchema } = wasteManagementTourSchemas;
-const logger = createSdkLogger({ component: 'waste-management-auth-runtime', level: 'info' });
-
-const toErrorLogFields = (error: unknown) =>
-  typeof error === 'object' && error !== null
-    ? {
-        error_type: error instanceof Error ? error.constructor.name : typeof error,
-        error_message: error instanceof Error ? error.message : String(error),
-        error_name: 'name' in error && typeof error.name === 'string' ? error.name : undefined,
-        error_code: 'code' in error && typeof error.code === 'string' ? error.code : undefined,
-        error_constraint:
-          'constraint' in error && typeof error.constraint === 'string' ? error.constraint : undefined,
-        error_detail: 'detail' in error && typeof error.detail === 'string' ? error.detail : undefined,
-        error_table: 'table' in error && typeof error.table === 'string' ? error.table : undefined,
-      }
-    : {
-        error_type: error instanceof Error ? error.constructor.name : typeof error,
-        error_message: error instanceof Error ? error.message : String(error),
-        error_name: undefined,
-        error_code: undefined,
-        error_constraint: undefined,
-        error_detail: undefined,
-        error_table: undefined,
-      };
 
 export const wasteManagementTourHandlers = {
   createWasteManagementTourInternal: async (
@@ -228,41 +208,19 @@ export const wasteManagementTourHandlers = {
     }
 
     try {
-      logger.info('waste_tour_delete_requested', {
-        operation: 'delete_waste_tour',
-        tour_id: tourId,
-        request_method: request.method,
-        request_url: request.url,
-        ...buildLogContext({ kind: 'instance', instanceId }, { includeTraceId: true }),
-      });
+      logWasteTourDeleteRequested(request, instanceId, tourId);
 
       const existing = await requireDeps(deps.loadWasteTourById, 'loadWasteTourById')(instanceId, tourId);
       if (!existing) {
         return createApiError(404, 'not_found', 'Die Waste-Tour wurde nicht gefunden.', requestId);
       }
 
-      logger.info('waste_tour_delete_existing_loaded', {
-        operation: 'delete_waste_tour',
-        tour_id: tourId,
-        tour_name: existing.name,
-        recurrence: existing.recurrence,
-        location_count: existing.locationCount,
-        waste_fraction_ids: existing.wasteFractionIds,
-        ...buildLogContext({ kind: 'instance', instanceId }, { includeTraceId: true }),
-      });
+      logWasteTourDeleteLoaded(instanceId, tourId, existing);
 
       await deleteWasteTourDependencies({ deps, instanceId, tourId });
-      logger.info('waste_tour_delete_final_delete_started', {
-        operation: 'delete_waste_tour',
-        tour_id: tourId,
-        ...buildLogContext({ kind: 'instance', instanceId }, { includeTraceId: true }),
-      });
+      logWasteTourDeleteFinalDelete(instanceId, tourId, 'started');
       await requireDeps(deps.deleteWasteTour, 'deleteWasteTour')(instanceId, tourId);
-      logger.info('waste_tour_delete_final_delete_completed', {
-        operation: 'delete_waste_tour',
-        tour_id: tourId,
-        ...buildLogContext({ kind: 'instance', instanceId }, { includeTraceId: true }),
-      });
+      logWasteTourDeleteFinalDelete(instanceId, tourId, 'completed');
 
       await emitWasteAuditEvent({
         deps,
@@ -283,28 +241,14 @@ export const wasteManagementTourHandlers = {
       if (error instanceof Error && error.message.startsWith('missing_dependency:')) {
         throw error;
       }
-      const isConflict = typeof error === 'object' && error !== null && 'code' in error && (error as { code?: string }).code === '23503';
-      logger.error('waste_tour_delete_failed', {
-        operation: 'delete_waste_tour',
-        tour_id: tourId,
-        is_conflict: isConflict,
-        ...toErrorLogFields(error),
-        ...buildLogContext({ kind: 'instance', instanceId }, { includeTraceId: true }),
-      });
-      await emitWasteAuditEvent({
+      return createWasteTourDeleteErrorResponse({
         deps,
         ctx,
         instanceId,
-        actionId: 'waste-management.tour.deleted',
-        result: 'failure',
-        reasonCode: isConflict ? 'conflict' : 'database_unavailable',
-        resourceType: 'waste_tour',
-        resourceId: tourId,
+        requestId,
+        tourId,
+        error,
       });
-      await updateWasteVisibleStatus(deps, instanceId, 'revalidate');
-      return isConflict
-        ? createApiError(409, 'invalid_request', 'Die Waste-Tour kann wegen bestehender Zuordnungen nicht gelöscht werden.', requestId)
-        : createApiError(503, 'database_unavailable', 'Die Waste-Tour konnte nicht gelöscht werden.', requestId);
     }
   },
 };
