+++ packages/plugin-waste-management/src/waste-management.tours.tour-mutations.ts
@@ -2,7 +2,6 @@ import { startTransition, type FormEvent } from 'react';
 import type { WasteTourRecord } from '@sva/plugin-sdk';
 
 import {
-  appendWasteManagementDebugLog,
   createWasteManagementTour,
   createWasteManagementLocationTourPickupDate,
   deleteWasteManagementLocationTourPickupDate,
@@ -11,6 +10,12 @@ import {
   updateWasteManagementTour,
 } from './waste-management.api.js';
 import { resolveApiErrorCode } from './waste-management.page.support.js';
+import {
+  logWasteTourDeleteError,
+  logWasteTourDeleteStart,
+  logWasteTourDeleteSuccess,
+} from './waste-management.tours.delete-debug.js';
+import { setTourDeleteErrorMessage } from './waste-management.tours.messages.js';
 import {
   createTourDateLocationAssignmentKey,
   isCustomDatesRecurrence,
@@ -24,19 +29,6 @@ import type { WasteToursState } from './use-waste-tours-state.js';
 type Translate = (key: string, variables?: Readonly<Record<string, string | number>>) => string;
 
 type WasteToursSubmissionContext = { readonly state: WasteToursState; readonly pt: Translate; readonly loadOverview: (active?: boolean) => Promise<void> };
-
-const setDeleteErrorMessage = (state: WasteToursState, pt: Translate, error: unknown) => {
-  const code = resolveApiErrorCode(error);
-  state.setMessage({
-    kind: 'error',
-    text:
-      code === 'forbidden'
-        ? pt('tours.messages.deleteForbidden')
-        : code === 'invalid_request'
-          ? pt('tours.messages.deleteConflict')
-          : pt('tours.messages.deleteError'),
-  });
-};
 
 const validateTourAssignments = (state: WasteToursState, pt: Translate) => {
   const activeDates = new Set(state.tourForm.customDates.map((entry) => entry.date));
@@ -186,19 +178,9 @@ const createDeleteTourHandler = ({ state, pt, loadOverview }: WasteToursSubmissi
   state.setMessage(null);
   state.setLastOutcome(null);
   try {
-    appendWasteManagementDebugLog({
-      scope: 'tour-delete',
-      phase: 'start',
-      tourId: tour.id,
-      tourName: tour.name,
-    });
+    logWasteTourDeleteStart(tour);
     await deleteWasteManagementTour(tour.id);
-    appendWasteManagementDebugLog({
-      scope: 'tour-delete',
-      phase: 'success',
-      tourId: tour.id,
-      tourName: tour.name,
-    });
+    logWasteTourDeleteSuccess(tour);
     await loadOverview(true);
     startTransition(() => {
       state.setMessage({
@@ -207,15 +189,8 @@ const createDeleteTourHandler = ({ state, pt, loadOverview }: WasteToursSubmissi
       });
     });
   } catch (saveError) {
-    appendWasteManagementDebugLog({
-      scope: 'tour-delete',
-      phase: 'error',
-      tourId: tour.id,
-      tourName: tour.name,
-      errorCode: resolveApiErrorCode(saveError) ?? undefined,
-      errorMessage: saveError instanceof Error ? saveError.message : String(saveError),
-    });
-    setDeleteErrorMessage(state, pt, saveError);
+    logWasteTourDeleteError(tour, saveError);
+    setTourDeleteErrorMessage(state, pt, saveError);
   } finally {
     state.setSaving(false);
   }
@@ -259,9 +234,9 @@ const createDeleteToursHandler = ({ state, pt, loadOverview }: WasteToursSubmiss
       return;
     }
 
-    setDeleteErrorMessage(state, pt, failedResults[0]?.reason);
+    setTourDeleteErrorMessage(state, pt, failedResults[0]?.reason);
   } catch (saveError) {
-    setDeleteErrorMessage(state, pt, saveError);
+    setTourDeleteErrorMessage(state, pt, saveError);
   } finally {
     state.setSaving(false);
   }
